Ensure rolltable displays on expand
This commit modifies the rolltable component to display the
actual table if it is not collapsed. It also ensures that
the result logic for random and selected entries works.

This commit moves the selection logic required in both the
parent rolltable component and the csvtable component out
into a referable file.

One known problem with this implementation is that the random
entry selector on the parent component cannot be clicked
unless the rolltable is expanded; this will need to be fixed
in a future commit.