Fix rolltable randomize and collapse
This commit modifies the logic for both the collapse
and randomize features of the rolltable component to
check first if the fileData state exists.

This fixes a problem in the prior commit where the
rolltable could not be used to select a random entry
unless it was displayed; instead of removing the table
from the DOM, it gives it the hidden class if collapsed.

This also changes the default behavior of the rolltable
card, displaying the `cannot generate...` warning if and
only if the CSV could not be parsed. Otherwise, it will
currently display nothing in that space until a result
is selected (intentionally or randomly).

A future commit could make it display a random result
by default instead, which might be more useful. Possibly,
that behavior should be toggleable.