Add debug and upload flags to rolltable
This commit adds flags to toggle the behavior
of the rolltable component to allow an author
to specify whether or not to display debug
info (valuable during authoring, not very
useful to end-readers) and to allow a custom
upload (possibly overwriting the default list
currently in use - ie overwrite the items list
or traits or something.

Both of these behaviors were previously turned
on by default for all rolltables. They are now
turned off by default.

This commit also updates the example page.